feat: add dart function to handle Sign in with Apple
The native Sign in with Apple flow uses the iOS' native OS to authenticate and generate an authorization code. This is preferable to the traditional OAuth2 flow because no new webview is required, and the user can make use of touch ID or Face ID to authenticate.

In order to finish the auth flow, the code needs to be exchanged for an ID token server-side. This function does that, and also creates a new user if one with the same email or id doesn't already exist.